Reduce SDS serial reading hangs
There was another softwareserial behavior change in last arduino core
update that extends the timeout behavior. as long as bytes keep flowing,
it will keep reading, which isn't what we expected.

So we need to go back to a state machine and read only when enough
has been buffered to stay responsive.